move collectLinks and validateLinks under helpers.ts
